Make base platform have sensors. Simplifies the inheritance considerably, without excluding any use cases.
